+++ src/renderer/global.d.ts
@@ -1,9 +1,16 @@
+/// <reference types="vite/client" />
+
 import type { CaptyAPI } from "../preload/index";
 
 declare global {
   interface Window {
     capty: CaptyAPI;
   }
+}
+
+declare module "*.onnx?url" {
+  const src: string;
+  export default src;
 }
 
 export {};